feat: all quelity checks now display the same error message structure
